Запрос на получение ссылки для аватара ProfileImageService
