+++ ImageFeed/Constants.swift
@@ -9,8 +9,8 @@ import Foundation
 
 enum Constants {
     
-    static let accessKey = "zCjCSr8jtSOUzOe1wWE7rO2z-9jBhpF9dkQoRiNsugM"
-    static let secretKey = "62iefqS86uTN7dN18YrvJL25U3EfMt0tzwkcbJI9mzo"
+    static let accessKey = "WdVo_zQdgh5P0B5aolxwPrpVbMxWfRBJaYdup--Y8UY"
+    static let secretKey = "Ja2fOeQtq8JE58AcV_EPNgN0qE-JgLUNZuET6tEJElI"
     static let redirectURI = "urn:ietf:wg:oauth:2.0:oob"
     static let accessScope = "public+read_user+write_likes"
