+++ ImageFeed/Services/OAuth2Service.swift
@@ -1,0 +1,76 @@
+//
+//  OAuth2Service.swift
+//  ImageFeed
+//
+//  Created by Наиль on 27/08/23.
+//
+
+import Foundation
+
+final class OAuth2Service {
+    
+    static let shared = OAuth2Service()
+    private let urlSession = URLSession.shared
+    
+    private var currentTask: URLSessionTask?
+    private var lastCode: String?
+    
+    private (set) var authToken: String? {
+        get { return OAuth2TokenStorage().token }
+        set { OAuth2TokenStorage().token = newValue ?? "" }
+    }
+    
+    func fetchAuthToken(code: String, completion: @escaping (Result<String, Error>) -> Void) {
+        
+        if lastCode == code { return }
+        
+        currentTask?.cancel()
+        lastCode = code
+        
+        let request = authTokenRequest(code: code)
+        
+        currentTask = urlSession.object(for: request) {[weak self] (result: Result<OAuthTokenResponseBody, Error>)  in
+            self?.currentTask = nil
+           
+            guard let self = self else {return }
+            
+            switch result {
+            case .success(let body):
+                let authToken = body.accessToken
+                self.authToken = authToken
+                completion(.success(authToken))
+            case .failure(let error):
+                completion(.failure(error))
+            }
+        }
+    }
+}
+
+extension OAuth2Service {
+    
+    private struct OAuthTokenResponseBody: Codable {
+        let accessToken: String
+        let tokenType: String
+        let scope: String
+        let createdAt: Int
+        
+        private enum CodingKeys: String, CodingKey {
+            case accessToken = "access_token"
+            case tokenType = "token_type"
+            case scope
+            case createdAt = "created_at"
+        }
+    }
+    
+    private func authTokenRequest(code: String) -> URLRequest {
+        URLRequest.makeHTTPRequest(
+            path: "/oauth/token"
+            + "?client_id=\(Constants.accessKey)"
+            + "&&client_secret=\(Constants.secretKey)"
+            + "&&redirect_uri=\(Constants.redirectURI)"
+            + "&&code=\(code)"
+            + "&&grant_type=authorization_code",
+            httpMethod: "POST",
+            baseUrl:  URL(string: "https://unsplash.com")!)
+    }
+}
