+++ ImageFeed/SplashView/SplashViewController.swift
@@ -12,9 +12,15 @@ import ProgressHUD
 class SplashViewController: UIViewController {
     
     private let showAuthenticationScreenSegueIdentifier = "ShowAuthenticationScreen"
-    private let oAuth2Service = OAuth2Service()
+    
+    private let oAuth2Service = OAuth2Service.shared
     private let oAuth2TokenStorage = OAuth2TokenStorage()
-    private let profileService = ProfileService.shared
+    
+    private let profileService = ProfileInfoService.shared
+    private let profileImageService = ProfileImageService.shared
+    
+    override func viewDidLoad() {
+    }
     
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
@@ -24,7 +30,6 @@ class SplashViewController: UIViewController {
     private func isAuth(){
         if let token = oAuth2TokenStorage.token {
             fetchProfile(token: token)
-            switchToTabBarController()
         } else {
             performSegue(withIdentifier: showAuthenticationScreenSegueIdentifier, sender: nil)
         }
@@ -61,18 +66,20 @@ extension SplashViewController: AuthViewControllerDelegate {
     func authViewController(_ vc: AuthViewController, didAuthenticateWithCode code: String) {
         dismiss(animated: true) { [weak self] in
             guard let self = self else { return }
-            UIBlockingProgressHUD.show()
             self.fetchOAuthToken(code: code)
         }
     }
     private func fetchOAuthToken(code: String) {
+        
+        UIBlockingProgressHUD.show()
+        
         oAuth2Service.fetchAuthToken(code: code) { [weak self] result in
-            guard let self = self else {return }
+            guard let self = self else { return }
             switch result {
             case .success(let token):
                 self.fetchProfile(token: token)
             case .failure:
-                ProgressHUD.dismiss()
+                UIBlockingProgressHUD.dismiss()
                 break
             }
         }
@@ -82,13 +89,25 @@ extension SplashViewController: AuthViewControllerDelegate {
         profileService.fetchProfile(token) { [weak self] result in
             guard let self = self else { return }
             switch result {
-            case .success:
-                UIBlockingProgressHUD.dismiss()
-                self.switchToTabBarController()
+            case .success(let profile):
+                self.fetchProfileImage(username: profile.userName, token: token)
             case .failure:
                 UIBlockingProgressHUD.dismiss()
                 break
             }
+        }
+    }
+    
+    private func fetchProfileImage(username: String, token: String) {
+        profileImageService.fetchProfileImageURL(username: username, token: token) { [weak self] result in
+            guard let self = self else { return }
+            switch result {
+            case .success:
+                self.switchToTabBarController()
+            case .failure:
+                break
+            }
+            UIBlockingProgressHUD.dismiss()
         }
     }
 }
